% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autocrunch.R
\name{autocruncher}
\alias{autocruncher}
\title{Quickly find knots that will cause issues for soap-film smoother
Setting up a soap film smoother is often a hard and frustrating process.
This function checks that all knots, not just one knot are within the boundary
and futher enough away from the boundary to effectively use a soap-film smoother}
\usage{
autocruncher(bnd, knots, k = 10, nmax = 200, xname = "x", yname = "y")
}
\arguments{
\item{bnd}{A list with sub-lists that will be the boundary the soap-film smoother
will smooth in. Coordinates need to be in a reference system that has meaningful units.
For example, coordinates could be in UTMs as UTMs rely on metres.}

\item{knots}{A dataframe with two columns that are the coordinates of the knots
that are to be supplied to the soap-film smoother.
Coordinates need to be be in meaniful units. For example, coordinates could be
in UTMs as UTMs rely on metres.}

\item{k}{If supplying \code{k} argument to \code{mgcv::gam()}, provide that
same value to \code{autocruncher()}.}

\item{nmax}{If supplying \code{nmax} argument to \code{mgcv::gam()}, provide that
same value to \code{autocruncher()}.}

\item{xname}{Column name of x coordinate for the knots object}

\item{yname}{Column name of y coordinate for the knots object}
}
\value{
A vector of the indices of all of the knots that either fall outside
the boundary or too close to the boundary and will cause for the soap-film
smoother to error.
}
\description{
Quickly find knots that will cause issues for soap-film smoother
Setting up a soap film smoother is often a hard and frustrating process.
This function checks that all knots, not just one knot are within the boundary
and futher enough away from the boundary to effectively use a soap-film smoother
}
\examples{
# crunch_ind <- autocruncher(fsb, knots, k=30, xname="v", yname="w"), will add more
crunch_ind <- autocruncher(sissabagama_lake_ls, sissabagama_lake_knots)

}
\references{
This function is based on soap.r in {mgcv}, Copyright Simon Wood 2006-2012.
Bugs fixes added by David L Miller, 2019
}
