% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{overlap_polygons}
\alias{overlap_polygons}
\title{The boundary list of example overlapping polygons}
\format{
A list containing 3 sub-lists with each sub-list containing two vectors, x and y.
\describe{
\item{x}{The x coordinates of each polygon, name needs to be x}
\item{y}{The y coordinates of each poloygon, name needs to be y}
}
}
\usage{
overlap_polygons
}
\description{
A list with sub-lists containing the x and y coordinates of 3 overlapping polygons.
This list of boundaries is used as an example to demonstrate how soap_check functions
when polygons overlap.
}
\keyword{datasets}
