% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sissabagama_bath}
\alias{sissabagama_bath}
\title{Sampled bathymetry points for Sissabagama Lake}
\format{
A dataframe containing 445 rows and 3 variables:
\describe{
\item{x}{The longitude of each sample}
\item{y}{The latitude of each sample}
\item{depth}{The latitude of each sample}
}
}
\usage{
sissabagama_bath
}
\description{
The latitude and longitude of sampled depth derived from bathymetric map from the
\href{https://apps.dnr.wi.gov/lakes/lakepages/LakeDetail.aspx?wbic=2393500}{Wisconsin DNR} for
Sissabagama Lake in northern Wisconsin.
}
\keyword{datasets}
