% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sissabagama_lake_grid_knots}
\alias{sissabagama_lake_grid_knots}
\title{Knots for Sissabagama Lake created through using a grid}
\format{
A dataframe containing 1,125 rows and 2 variables:
\describe{
\item{lon}{The longitude of each knot}
\item{lat}{The latitude of each knot}
}
}
\usage{
sissabagama_lake_grid_knots
}
\description{
The latitude and longitude of knots evenly spaced 200 m away from each other
covering the boundary box of Sissabagama lake. Knots included are outside the
boundary of the lake.
}
\keyword{datasets}
