% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sissabagama_lake_knots}
\alias{sissabagama_lake_knots}
\title{Knots for Sissabagama Lake}
\format{
A dataframe containing 420 rows and 2 variables:
\describe{
\item{lon}{The longitude of each knot}
\item{lat}{The latitude of each knot}
}
}
\usage{
sissabagama_lake_knots
}
\description{
The latitude and longitude of knots evenly spaced 200 m away from each other inside
Sissabagama Lake in northern Wisconsin.
}
\keyword{datasets}
