% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sissabagama_lake_ls}
\alias{sissabagama_lake_ls}
\title{The boundary list of Sissabagama Lake}
\format{
A list containing 5 sub-lists with each sub-list containing two vectors, x and y.
\describe{
\item{x}{The longitude of each polygon, name needs to be x}
\item{y}{The latitude of each poloygon, name needs to be y}
}
}
\usage{
sissabagama_lake_ls
}
\description{
A list with sub-lists containing the latitude and longitude of each polygon that
make up Sissabagama Lake in northern Wisconsin.
}
\keyword{datasets}
