% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sissabagama_lake_sf}
\alias{sissabagama_lake_sf}
\title{Sissabagama Lake located in northern Wisconsin}
\format{
Simple feature collection with 1 features and 23 fields:
\describe{
\item{geometry}{POLYGON}
}
}
\usage{
sissabagama_lake_sf
}
\description{
Sissabagama Lake located in northern Wisconsin. The original shapefile was derived from
the Wisconsin DNR's watershed database. Latitude and longitude are in UTM zone 15N.
}
\keyword{datasets}
