% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/soap_check.R
\name{soap_check}
\alias{soap_check}
\title{Check if boundary, knots, and data can be modeled using a soap-film smoother}
\usage{
soap_check(
  bnd,
  knots = NULL,
  data = NULL,
  plot = TRUE,
  tol = sqrt(.Machine$double.eps),
  x_name = "x",
  y_name = "y"
)
}
\arguments{
\item{bnd}{A list with sub-lists that will be the boundary the soap-film smoother
will smooth in. Coordinates need to be in meaningful units.
For example, coordinates could be in UTMs as UTMs rely on metres.}

\item{knots}{A dataframe with two columns that are the coordinates of the knots
that are to be supplied to the soap-film smoother. Coordinates need to be in
meaningful units. For example, coordinates could be in UTMs as UTMs rely on metres.}

\item{data}{A dataframe with two columns that are the coordinates of the data
that are to be supplied to the soap-film smoother. Coordinates need to be in
meaningful units. For example, coordinates could be in UTMs as UTMs rely on metres.}

\item{tol}{Tolerance value to check if boundaries are complete polygons.
Sometimes tolerance needs to be increased (e.g., tol = 1e-6)}
}
\value{
TRUE or FALSE depending on whether the boundary will be able to used
by a soap-film smoother. Addtionally if supplying knots and/or data. It will warn
the user which knots and/or data fall too close or outside the boundary.
}
\description{
Setting up a soap film smoother is often a hard and frustrating process.
This function checks that the boundary, knots, and data that you feed to
a soap-film smoother are “correct”. The function will plot  the boundary,
knots and data that are trying to be modeled to ensure that they are are
appropriate
}
\examples{
# library(mgcv)
# fsb <- list(fs.boundary())
# soap_check(fsb)

soap_check(sissabagama_lake_ls)

}
